requirements.txt
pygame==1.9.4
pyyaml